% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCFserver.R
\name{VCFserver}
\alias{VCFserver}
\title{create the server functions}
\usage{
VCFserver(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
create the server functions
}
