% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{startVCFshiny}
\alias{startVCFshiny}
\title{Start the VCFshiny shiny App}
\usage{
startVCFshiny(
  launch.browser = TRUE,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\description{
Start the VCFshiny shiny App
}
